refactor web content import, add source
