+++ .gitignore
@@ -47,3 +47,4 @@ data/BkM/raw/20221221_102750_object.jsonl
 data/ULAN/json/ulanArtists.jsonl
 data/ULAN/json/ulanCorporateBodies.jsonl
 data/ULAN/raw/*
+data/BrooklynMuseum/collections.jsonl
